Fix #661 : Log download timeout is too short
Issue:

There is a hard coded timeout of 10 seconds from the start of the download of execution logs to the end. This is too short for longer logs and for connections that have longer latencies.

The suggested action item to refresh the page won't help either.

"
The log is taking a long time to finish loading. Azkaban has stopped loading them. Please click Refresh to restart the load.
"

Fix:
Removed the limit which is set in the client side Javascript.

Test:
Manually shortened the timeout value to 10 milliseconds to reproduce the behavior. Applied the fix and verified that the log displayed correctly.