Track hits to Dr. Elephant which are coming from azkaban. This change enables us to track the links originating from Dr. Elephant in piwik dashboard.
Notes:
Azkaban is https and the outgoing link to Dr. Elephant is http and hence by policy "referrer" header won't be set by the browser.
An alternate way will be to set the meta tag in the relevant page. Going by the above approach since it doesn't require any code change.