Moving Azkaban Hadoop Security code from azkaban-plugins to this repo
The current changes consolidates azkaban-plugins/hadoopsecuritymanager-yarn and azkaban-plugins/hadoopsecuritymanager-common into a single submodule azkaban-hadoop-security-plugin in this repo. Following this commit, the Azkaban hadoop security code in the azkaban-plugins will be deprecated but will remain as is. However, future deployments of Azkaban will use this module as opposed to the one in azkaban-plugins.

Motivation: Moving the plugin code into the main repo helps in consolidation. The azkaban-plugins repo is not maintained well. It currently uses hardcoded and uploaded dependencies in the repo which is not a good model for managing artifacts. The new module works with gradle and pulls dependencies using ivys. This also allows the hadoop security code to managed much easily and more closely with the entire Azkaban project.

Code Changes: Very minimal code changes have been made to make the code compile. The changes are mostly wrapping toString() calls on byte arrays to Arrays.toString(). This is flagged as an error by the errorprone library.