Cosmetic: Renamed azkaban-webserver, azkaban-execserver and azkaban-soloserver modules (#737)
Renamed azkaban-webserver, azkaban-execserver and azkaban-soloserver modules to azkaban-web-server, azkaban-exec-server and azkaban-solo-server respectively.

Motivation: Previously, the names of Azkaban artifacts were manually fed. Thus, despite the name of the project being azkaban-webserver, it would still create artifacts with name azkaban-web-server which makes it inconsistent with the project name. This renaming would keep the name consistent with the previous artifacts.

Updated build scripts to point to the new names of these modules.

The build passes successfully with all tests. Artifacts are generated with the correct names.