Throw detailed exceptions and better error messages in ProjectManagerResource (#733)
* Throw exception on deploy action fail. 
* Added unit tests to make sure this behaviour happens (#725)
* Made deploy action code raise RestLi exceptions on failure. (#729)
This allows us to choose the specific HTTP status code when operations
fail. The status codes for this action have been updated to return 400 on failure.

* Made checkReport package-private for encapsulation. (#729)
Also moved test file for this into a different package. Condensed some
of its logic for readability, and added a new test to ensure messages
are getting through on error.

* Resolved file name change problem (#725)

* Made DeployProjectTest more compact, and renamed checkReport to
checkReports. (#725)