Enable using log4j.properties file in conf directory (#768)
Right now, Azkaban searches for a log4j.properties file baked inside of
a jar file. This change makes the Azkaban startup script search for this
file inside of the conf directory and use it if it exists. If it does
not exist, then we will revert to using the properties file found in the
jar by default.