Option to create Kafka Appender for user job logs (#778)
JobRunner now checks for properties in jobs to check if Kafka logging is
enabled. If it is, then it will attempt to create a Kafka appender. If
that fails, it will report the failure but not crash the application.

Job properties added (with example values):
azkaban.job.kafkalogging.enabled=true
azkaban.job.kafkalogging.brokerlist=localhost:9092
azkaban.job.kafkalogging.topic=azkaban-web-server