Deleted Azkaban config files from repo (#805)
Currently, the deleted files are being written into Azkaban jars. These
do not belong here, and can cause hidden configurations from being used.

Consequently, azkaban-solo-server will not work as it relies on these
files, so solo-server was given its own conf directory.

Also, h2 was added as a runtime dependency for solo-server, which
happened to be missing for the past little while.