Executor to use any port if not configured. Added REST APIs to activate / deactivate executor (#813)
When an executor server is launched, it uses the configured port or falls back to a default port. This behavior has been modified such that when the executor is not configured to a specific port it defaults to any available port that it can find on the machine. The advantage is that the user doesn't need to worry about port collisions when he/she doesn't need a specific port.

Also, The executor now automatically adds an entry about itself to the DB post launch. Also, it exposes the following REST APIs to allow users to activate / deactivate itself
  example:
  http://localhost:1234/executor?action=activate
  http://localhost:1234/executor?action=deactivate

Other changes:
* Remove public signature from interface methods
* Write port to executor.port file in working directory
* Refactored some constants which were duplicated across azkaban-common and azkaban-exec-server