Create PatternLayout that escapes troublesome characters (#821)
When we use log4j to send a JSON message the official PatternLayout will
send corrupt JSON if certain characters are present. We use this
PatternLayoutEscaped class as a thin interface around PatternLayout to
escape all these characters.

An example use case of this is if we are sending JSON with a
KafkaAppender. Multi lined messages result in corrupt JSON, requiring
that the new lines be escaped before being appended.

So far we know we need to escape backslashes, tabs, newlines and quotes.
Unit tests were added for these cases.