Option to create Kafka Appender for user job logs (#797)
* Option to create Kafka Appender for user job logs (#778)

JobRunner now checks for properties in jobs to check if Kafka logging is
enabled. If it is, then it will attempt to create a Kafka appender. If
that fails, it will report the failure but not crash the application.

User side:
A basic example usage of this would be to drop a properties file (named maybe like kafka.properties) with the following lines into the root directory of your project:
azkaban.job.logging.kafka.enable=true

This property should cause all the jobs in your project to log to Kafka, using a producer provided by the executor server. Note that the executor server can override this property and refuse to provide this producer.

The property azkaban.job.logging.kafka.enable is by default false, so if left alone, the default behavior is to not log to Kafka.

Server admin side:
There are two properties to configure the Kafka producer that would be used as a logger for user jobs and they can be set in azkaban.properties. They are:
azkaban.server.logging.kafka.brokerList
azkaban.server.logging.kafka.topic

If any of these two properties are missing, then the executor server will refuse to attempt to make the producer. This can doubly act as a killswitch if the server admins decide they don't want anyone logging to Kafka, or don't have a Kafka cluster set up.

If during the creation of the Kafka logger, something wrong happens, then we will fail silently (not crash the app), and report the failure in the flow log.

* Refactored Constants.java into separate files and added new constants

Added some configuration strings into the Constants.java file. These
configurations specify properties to manage user job Kafka logging, as
well as job and flow level properties.

* Added server props object to job and flow runners
Changed job and flow runner objects to have a copy of the Azkaban
Executor server properties, so that when running flows we can activate
server wide configurations.

As a result, a bunch of tests had to be altered.