Use hard link in execution dir to prevent flow failure where executor is reading deleted project dirs. (#894)
Since executor is not aware of running flows of other executors, it's possible that one executor is deleting the project which others are running. Previously we use symlink in execution dir. Symlink will be invalidated and execution will thus fail when source project dir is deleted by other executors when multiple executors are running. In the case when two executors are downloading the same project, the logic will still work since execution dirs will be linked to one project dir and no deletion will take place for new projects. Please note that due to the restriction of hard link, project dir and execution dir have to be in the same partitions.