Retry the jobs on executor instead of killing them right away when executor host is running out of memory (#937)
Currently a job will be killed right away before actual execution if there's not enough memory. But as memory usage fluctuates, it's possible there will be enough memory later. Therefore we should retry the job if memory requirement is not met.

Testing:
Given the nature of this code change, it's difficult to add automated tests. For now we tested the change as by replacing SystemMemoryInfo.canSystemGrantMemory with a function which returns true or false based on given probability, then manually verified it's logging correctly and hitting the according code path.