Fixing js libraries version compatibility (#979)
Issue #975 reported a bug that the name of moment-timezone js is not
right, and gradle can not fetch it to distribution package.

Tried ligradle build, and figured out that the latest moment-timezone js 0.5.13 is not compatible iwht 0.5.5, since the js file built's name doesn't keep the same.

We use `^` to specify compatible js libraries (see details in https://docs.npmjs.com/misc/semver). In this change, we disable it to in order to secure repo.