Storage Stubs and Implementation + Guice Service Provider (#991)
This patch introduces a `ServiceProvider` interface powered by `Guice` which is also a new dependency for Azkaban. This was required to implement some of the storage implementation code.

Storage API changes
 - refactored some of the storage APIs
 - Implemented Local storage for Azkaban + unit test
 - Included Stub for Database Storage.

None of these code paths are invoked anywhere in the codebase at this time. So this should not affect any existing code. Just wanted to break up my changes into smaller chunks to make it easier to review.