Fix flaky metrics related tests (#1018)
Instead of using a reporter which is subject to the multi-thread
related timing issues, verify that the metrics registry has the
 correct information. We don't have to test that reporter works here.

Redesigned the tests to make the tests more explicit and more closely
reflect the actual usage pattern of the metrics related methods.
The existing shared test routines make strong assumptions about the
behavior of the functional methods passed to them. IMHO they make the
test code harder to read.

The meter related tests are a bit trickier to test since they
return double type.
The existing meter related tests only test the count not the actual
value reported in the AutoMetrics anyway.
Remove these tests for now until we have a better solution.

Removed the empty Executor Metrics test class.