Wiring get calls via StorageManager (#1019)
This changes delegates all get calls via the `StorageManager`. By `get` I mean the process of fetching a project file (blob) from storage. By default, this storage is the database. There are some hacks in the code which can be fixed later as we go.

There are some refactors in the executor codebase. This was done to easily fetch the `StorageManager` class via Guice.

The changes have been verified on solo server locally and also on executors separately.

Others
 - Changed catch block as per previous review comment
 - Updated todos with more detail