Fix slow BasicTimeCheckerTest
Issue:

The test sleep for 20 seconds and the comments are wrong.
When debugging the test, the test would fail since it is time sensitive.

Fix:
Switch to use Jodatime utility to get the current time and manipulate
 the time in the test.

It's better to use the java time api and use a Guice injected wrapper
 class to manipulate time.
 However given that BasicTimeChecker is scheduled to be deprecated
 soon. It's not worth investing in it now.