HDFS instance needs to be created post UGI login
Despite authorizing prior to any HDFS call, the code does not seem to work without an existing kinited session.
The problem was that the `FileSystem.get()` API needs to be called after a successful UGI login.

The Hadoop code caches the logged in user at the time of creating the object and this is not reflected in the
stack traces.

Fix: ensure UGI auth prior to creation of file system object

(cherry picked from commit 853f343)