Use RetryingHiveMetaStoreClient instead of HiveMetaStoreClient (#1235)
Azkaban currently uses `HiveMetaStoreClient` while communicating with the Hive Metastore ref class: `HadoopSecurityManager_H_2_0`
This client does not automatically retry/reconnect in case of transient failures, such as when a metastore is restarted. The RetryingHiveMetaStoreClient does automatically retry/reconnect in case of errors.

This change switches it to use the RetryingHiveMetaStoreClient (which is what the Hive CliDriver uses internally).

NOTE: This introduces a runtime dependency on `hive-exec` core jar. Currently, this is set as a compile only dependency in Azkaban but it actually requires this jar in the runtime and expects to get it from the `HIVE_HOME` environment variable.

Ref: https://github.com/linkedin/gobblin/issues/1506