fix endtime schedule (#1246)
An issue was introduced in #1110 that causing schedule end time is set to a past date whenever SLA changes are made.
It's caused by incompatibility of new code #1110 with old schedule metadata. SLA change triggers reloading updated schedules into memory. But instead of using a default schedule end time, which we set to some day in year 2050, it always mistakenly uses a past date as the expiring time of the schedule.
This fixes the issue by setting appropriate end time of the schedule.