Refactor: Guice injection of Azkaban Web Server dependencies
The current code uses the `SERVICE_PROVIDER` interface to fetch dependencies for the Web Server. Refactored the code to enable parameterized dependency injection from the construction which is the standard way of declaring dependencies.

Injection sanity tests already exist in the code. Tested solo server and it works as well.