Fix for failing to kill a job due to a race condition
Problem:

After a flow starts, kill the flow quickly.
The flow and the job will show as killed. However the job actually runs
to completion.

Analysis:

Jobrunner thread runs: azkaban.jobExecutor.utils.process
.AzkabanProcess#run

A jetty thread processes the kill command:

azkaban.jobExecutor.ProcessJob#cancel
azkaban.jobExecutor.utils.process.AzkabanProcess#softKill
azkaban.jobExecutor.utils.process.AzkabanProcess#checkStarted
here it throws exception, because the job process has not been created
yet at this point.

This exception is caught and ignored.
Users are informed that the kill action is completed and reflected in
the job and flow pages.

Fix:

Synchronize the killing thread and the jobrunner thread.
The jetty thread will wait for the job process to be created if needed
before killing the job process.

====

This is a fix built on the idea proposed in #1289
The previous fix #1253 for the same race condition is not complete.