Throwing error in assignUserDirOwnership in ProcessJob (#1334)
Updating PR #1325.

Instead of simply displaying an error message when assignUserDirOwnership fails, now it throws an Exception using the handleError method.

This is better because if the flow fails later due to unexpected permissions for the user then it would be hard to detect that the failure of assignUserDirOwnership is the root cause. Now the flow will fail quickly and be easier to diagnose.