fix access to input and output property files for python, ruby and script jobs
The LongArgJob was setting the environmental variables before the
input and output property files were created. This meant that the
process that was started did not have access to the input or output
property files.

These changes modify the LongArgJob so that the property files are
created before the environment variables are collected and configured
for the job.

As the environment was previously set during the initialization,
subclasses had a chance to add their own environmental variables
to the process builder before the process was started. To keep
this backwards compatible, any variables that have been added to
the environmental variable map before the job runs will be merged
to the final environment overriding any statically defined variables.

Fixes #319