Fixed "Center" button in graph view context menu (#1338)
The "centerNode" function calls getTransformToElement(), which was removed from the SVG spec and also from browsers such as Chrome and Firefox, so the "Center" context menu button that appears when right-clicking in the graph view doesn't work. Fortunately it doesn't seem to use this call for anything so simply removing the line makes the button work again.

Tested by injecting the modified svg-graph.js as a Greasemonkey script in Firefox 55.0.1, and locally modifying the script in the debugger in Chrome 60.0.3112.90.