One refactor for Executor operations from JdbcExecutorLoader class (#1357)
This code patch is a follow-up of #1345. Similar to that PR, I start moving executor related operations. The first step is to create a delegate class and move all related methods to the new class.

Note that I didn't change anything, but only move methods from one place to another place. After this PR, I will change the implementation inside, and add unit test. The benefits of this PR is to help team do review.