build: use a fixed google errorprone version
## Issue:
"The plugin adds an errorprone configuration that automatically uses the
 latest release of error-prone. You can override it to use a specific
 version "
see https://github.com/tbroyer/gradle-errorprone-plugin

It appears that by latest version it will use the last locally cached
version. This caused inconsistency in builds.

## Fix:
Specify a fixed version

Will update to the latest version after the internal Artifactory
is updated with the latest version.

## Test:
Deleted the gradle cache and run the build again and verified in the
console output that the specified version was downloaded.