rewrite assign executor jdbc methods (#1379)
Still, this PR is a follow-up of #1345. Last but not least, I'm moving assignExecutor and UnAssignExecutor to a new DAO file.

The reason I don't create a new test file:

Not usre how much in-memory h2 database consume resources. We could use only one test class (in-memory h2) to do testing.