Customize code style template. Update README
- Modify the style template to preserve line feeds in the Javadoc. This
makes it easier to preserve the Javadoc format in the source file. This
is especially helpful given this intellij issue
"Formatting doesn't apply to javadoc in modified files"
https://youtrack.jetbrains.com/issue/IDEA-178073

- Change the name of the style config file and the style name
to reflect that it is now a customized style for the project.
And it contains styles for other languages such as JavaScript as well.

- Update the README to reference the new style file. Make it clear that
the new style need to be activated after it is imported.

- Update the README to require saveAction plugin. We have finished
auto formatting all the source files.