Change style rules in the README file
The modification we made to the style configuration i.e. preserving line
feeds is still consistent with the Google style guide.

It's simpler to state we follow strictly the Google style.

Also added the link to the specific version of the config file our
configuration file is based on. It makes it easier to merge in new
changes to the Google style config in the future.