Enforce using Singleton Annotation rather than binding statement (#1428)
There still exists a few places where we bind class to singleton in guice module class. As discussed a few weeks ago, we decided to enforce using @Singleton annotation, rather than the binding statement. The benefit is to manage and view singleton class easily.

This code patch tries to fix/replace this issue.