Enable JobRunnerTest > testBasicRun (#1469)
Use the InteractiveTestJob type instead of the Java jobtype

The old JobRunner test runs a java job that launches an additional JVM / java process. That should be the responsibility of JavaProcessJobTest, and I think it already does that.

If this test case serves any purpose, I think it should be converted to use the InteractiveTestJob which is simple and quick and lives within the test process itself.