Creating DB Metrics (#1468)
The goal of this PR is to build metrics in azkaban-db module. The most important metrics are DB connection monitorings. Here's the steps to complete the task:

Create DBMetrics class;
make AzkabanDataSource Guiceable, in order to Inject DBMetrics. Adding Binding method in azkabanCommonModule to guice it;
Moving Common MetricsTestUtility method to Core module, so that db module can call it. Also, add test dependency gradle code.
After that, add 3 metrics to MysqlDataSource.