Fixing regression where JOB_OUTPUT_PROP_FILE was no longer writable by user process (#1504)
JOB_OUTPUT_PROP_FILE is a file used to pass properties between jobs. You can read about it here: http://azkaban.github.io/azkaban/docs/latest/#job-configuration

#1325 introduced a regression where JOB_OUTPUT_PROP_FILE became no longer writable due to an upgraded execute-as-user executable that stopped user processes from having the azkaban group permission. This file needs to be written to so properties can be passed from one job to another.

This fix assigns ownership of JOB_OUTPUT_PROP_FILE and JOB_PROP_FILE to <username>:azkaban to guarantee user ability to write to this file. It is uncertain whether or not JOB_PROP_FILE is intended to be written to, but I believe it's a good idea to do allow it to be in order to prevent regressions.

This is covered by an integration test that simply tries to write to the JOB_OUTPUT_PROP_FILE to confirm that it is doable. This is already in master of azkaban-auto-test.

This change will more drastically take effect when the execute-as-user executable is upgraded to its newest version.

This was tested on solo-server and holdem4.

I'll continue to explore JOB_OUTPUT_PROP_FILE and JOB_PROP_FILE usage and implementation so we can make sure these features are fully covered by tests and can be leveraged successfully by our users.