Show queued jobs in the execution list (#607)
The PR fixes a problem when a queued job would disappear from the job list view when its state changes to QUEUED.

Currently, you can access job logs for current and failed attempts in the Job List tab of an execution.

If some job failed and has retries, it goes into QUEUED state while sleeping before the next attempt.

If you want to see why any of the previous attempts failed, you can right-click the grey attempt bar
To see how long it's going to sleep, you can check the job log
Without this fix you can't do either of those if the job is sleeping, because it's hidden if the state is QUEUED.

Also if you have a job that hasn't yet started to run but is in QUEUED state (ie. blocked by pipeline), if you want to know why it's queued, you should check the job log. If you can't access the job log via Job List tab, this is not easily accomplished.