Build Quartz Job Factory to allow Quartz jobs easily inject dependencies (#1540)
We recently found Quartz jobs cannot easily inject a dependency. The previously proposal was to add dependencies into Context of Quartz Job. However, it is not recommended, the correct approach is to only inject String or very light weight objects into context. Besides, it increased more complexity for users to leverage Quartz. In this PR, we create a Quartz Job Factory, which is able to seamlessly provide any dependency inside web server.